+++ packages/nextjs/src/app-router/server/ClerkProvider.tsx
@@ -82,7 +82,6 @@ export async function ClerkProvider(
 
   let [shouldRunAsKeyless, runningWithClaimedKeys] = [false, false];
   if (canUseKeyless) {
-    // eslint-disable-next-line import/no-unresolved
     const locallyStorePublishableKey = await import('../../server/keyless-node.js')
       .then(mod => mod.safeParseClerkFile()?.publishableKey)
       .catch(() => undefined);
@@ -94,10 +93,11 @@ export async function ClerkProvider(
 
   if (shouldRunAsKeyless) {
     // NOTE: Create or read keys on every render. Usually this means only on hard refresh or hard navigations.
-    // eslint-disable-next-line import/no-unresolved
-    const newOrReadKeys = await import('../../server/keyless-node.js').then(mod => mod.createOrReadKeyless());
+
+    const newOrReadKeys = await import('../../server/keyless-node.js')
+      .then(mod => mod.createOrReadKeyless())
+      .catch(() => null);
     const { keylessLogger, createConfirmationMessage, createKeylessModeMessage } = await import(
-      // eslint-disable-next-line import/no-unresolved
       '../../server/keyless-log-cache.js'
     );
 
@@ -147,7 +147,6 @@ export async function ClerkProvider(
 
         output = clientProvider;
       } else {
-        // eslint-disable-next-line import/no-unresolved
         const KeylessCookieSync = await import('../client/keyless-cookie-sync.js').then(mod => mod.KeylessCookieSync);
 
         const headerStore = await headers();
@@ -172,6 +171,18 @@ export async function ClerkProvider(
 
         output = <KeylessCookieSync {...newOrReadKeys}>{clientProvider}</KeylessCookieSync>;
       }
+    } else {
+      // When case keyless should run, but keys are not available, then fallback to throwing for missing keys
+      output = (
+        <ClientClerkProvider
+          {...mergeNextClerkPropsWithEnv(rest)}
+          nonce={await generateNonce()}
+          initialState={await generateStatePromise()}
+          disableKeyless
+        >
+          {children}
+        </ClientClerkProvider>
+      );
     }
   }